fix: 알림 웹소켓 에러 핸들링
