+++ src/main/java/com/example/earthtalk/controller/NotificationController.java
@@ -5,7 +5,6 @@ import com.example.earthtalk.domain.notification.dto.request.RemoveTokenRequest;
 import com.example.earthtalk.domain.notification.dto.request.SaveTokenRequest;
 import com.example.earthtalk.domain.notification.dto.request.SendNotificationRequest;
 import com.example.earthtalk.domain.notification.dto.response.CheckTokenResponse;
-import com.example.earthtalk.domain.notification.dto.response.NotificationListResponse;
 import com.example.earthtalk.domain.notification.dto.response.NotificationListResponseWithUnreadCount;
 import com.example.earthtalk.domain.notification.service.FcmTokenService;
 import com.example.earthtalk.domain.notification.service.NotificationService;
@@ -14,12 +13,7 @@ import io.swagger.v3.oas.annotations.Operation;
 import io.swagger.v3.oas.annotations.tags.Tag;
 import lombok.RequiredArgsConstructor;
 import lombok.extern.slf4j.Slf4j;
-import org.springframework.data.domain.Page;
 import org.springframework.http.ResponseEntity;
-import org.springframework.messaging.handler.annotation.MessageMapping;
-import org.springframework.messaging.handler.annotation.Payload;
-import org.springframework.messaging.simp.SimpMessageHeaderAccessor;
-import org.springframework.messaging.simp.SimpMessagingTemplate;
 import org.springframework.web.bind.annotation.*;
 
 import java.util.List;
@@ -31,7 +25,6 @@ import java.util.List;
 @RequestMapping("/api/notifications")
 public class NotificationController {
 
-    private final SimpMessagingTemplate messagingTemplate;
     private final NotificationService notificationService;
     private final FcmTokenService fcmTokenService;
 
@@ -124,10 +117,4 @@ public class NotificationController {
         return ResponseEntity.ok(ApiResponse.createSuccess(null));
     }
 
-    @MessageMapping("/noti")
-    public void getSession(@Payload String tempId, SimpMessageHeaderAccessor accessor) {
-        String sessionId = accessor.getSessionId();
-        log.info("sessionId 매칭 tempId : {} | sessionId : {}", tempId, sessionId);
-        messagingTemplate.convertAndSend("/queue/handshake-" + tempId, sessionId);
-    }
 }
