--- src/main/java/com/example/earthtalk/domain/debate/component/StompPrincipal.java
@@ -1,17 +1,0 @@
-package com.example.earthtalk.domain.debate.component;
-
-import java.security.Principal;
-
-public class StompPrincipal implements Principal {
-
-    private final String name;
-
-    public StompPrincipal(String name) {
-        this.name = name;
-    }
-
-    @Override
-    public String getName() {
-        return name;
-    }
-}
